+++ src/Service/AppSync/src/Enum/DataSourceLevelMetricsConfig.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace AsyncAws\AppSync\Enum;
+
+final class DataSourceLevelMetricsConfig
+{
+    public const DISABLED = 'DISABLED';
+    public const ENABLED = 'ENABLED';
+
+    public static function exists(string $value): bool
+    {
+        return isset([
+            self::DISABLED => true,
+            self::ENABLED => true,
+        ][$value]);
+    }
+}